requirements.txt
flask
flask-socketio
flask-oauthlib
google-apps-meet
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-pubsub
google-cloud-speech
google-api-python-client
python-dotenv
flask-caching
flask_cors
pydub
pyttsx3
pyaudio
openai
langchain_core
langchain_community
sentence_transformers